Move appfile and experiments packages
